BM and RS Done

+++ KITE/Pages/ContentPages/GIRawMaterialUpload.aspx.cs
@@ -1,5 +1,4 @@
-﻿using CsvHelper;
-using KITE.Models;
+﻿using KITE.Models;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -12,7 +11,8 @@ namespace KITE.Pages.ContentPages
 {
     public partial class GIRawMaterialUpload : System.Web.UI.Page
     {
-        private List<GIRawMaterialViewModel> CsvDataList;
+        private List<GIRawMaterialWithConvertionViewModel> CsvDataList;
+
         protected void Page_Load(object sender, EventArgs e)
         {
             if (IsPostBack && CsvDataGridView.DataSource == null)
@@ -25,12 +25,13 @@ namespace KITE.Pages.ContentPages
         {
             errorLabel.Text = "";
             GIRawMaterialFunctionModel giRawMaterialFunction = new GIRawMaterialFunctionModel();
-            Tuple<Exception, string, List<GIRawMaterialViewModel>> fileResult = giRawMaterialFunction.GIRawMaterialReadCsvFile(fileUpload);
+
+            Tuple<Exception, string, List<GIRawMaterialWithConvertionViewModel>> fileResult = giRawMaterialFunction.GIRawMaterialReadCsvFile(fileUpload);
             if (fileResult.Item1.Message != "null")
             {
                 UtilityModel errorHandler = new UtilityModel();
                 Exception loadCsvException = new Exception("Terdapat masalah ketika mensubmit file csv. Mohon untuk cek kembali apakah data file csv cocok dengan format upload GI Raw Material.<br/> Detail : " + fileResult.Item1.Message);
-                errorHandler.UploadCsvErrorHandler(loadCsvException, CsvDataGridView, errorLabel);
+                errorHandler.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
             }
             else if (fileResult.Item2 != null && fileResult.Item3 != null)
             {
@@ -116,7 +117,7 @@ namespace KITE.Pages.ContentPages
             {
                 UtilityModel errorHandler = new UtilityModel();
                 Exception loadCsvException = new Exception("Terdapat masalah ketika mau membuka halaman " + txtGoToPage.Text.Trim() + ". Detail : " + ex.Message);
-                errorHandler.UploadCsvErrorHandler(loadCsvException, CsvDataGridView, errorLabel);
+                errorHandler.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
             }
         }
 
@@ -138,7 +139,7 @@ namespace KITE.Pages.ContentPages
             {
                 UtilityModel errorHandler = new UtilityModel();
                 Exception loadCsvException = new Exception("Terdapat masalah ketika tombol previous page berjalan. Detail : " + ex.Message);
-                errorHandler.UploadCsvErrorHandler(loadCsvException, CsvDataGridView, errorLabel);
+                errorHandler.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
             }
         }
 
@@ -161,7 +162,7 @@ namespace KITE.Pages.ContentPages
             {
                 UtilityModel errorHandler = new UtilityModel();
                 Exception loadCsvException = new Exception("Terdapat masalah ketika tombol next page berjalan. Detail : " + ex.Message);
-                errorHandler.UploadCsvErrorHandler(loadCsvException, CsvDataGridView, errorLabel);
+                errorHandler.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
             }
         }
 
@@ -175,20 +176,20 @@ namespace KITE.Pages.ContentPages
         {
             try
             {
-                // Load your CSV data here
-                ReadCsvModel readCsv = new ReadCsvModel();
-                using (CsvReader csvData = readCsv.ReadCsvFile((string)Session["FilePath"], ";"))
+                Tuple<object, Exception> readCsvResult = new ReadCsvModel().ReadCsvFunction("giRawMaterial", (string)Session["FilePath"], ";", ConfigurationManager.ConnectionStrings["DefaultConnection"].ConnectionString);
+                if (readCsvResult.Item2.Message != "null")
                 {
-                    Tuple<object, Exception> uomConvertionObject = new ReadCsvModel().UomConvertion(csvData, "giRawMaterial", ConfigurationManager.ConnectionStrings["DefaultConnection"].ConnectionString);
-                    CsvDataList = (List<GIRawMaterialViewModel>)uomConvertionObject.Item1;
-                    csvData.Dispose();
+                    UtilityModel errorHandler = new UtilityModel();
+                    Exception loadCsvException = new Exception("Terdapat masalah ketika memuat file csv. Mohon untuk cek kembali apakah data file csv cocok dengan format upload McFrame process.<br/> Detail : " + readCsvResult.Item2.Message);
+                    errorHandler.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
                 }
+                CsvDataList = (List<GIRawMaterialWithConvertionViewModel>)readCsvResult.Item1;
             }
             catch (Exception ex)
             {
                 UtilityModel errorHandler = new UtilityModel();
                 Exception loadCsvException = new Exception("Terdapat masalah ketika memuat file csv. Mohon untuk cek kembali apakah data file csv cocok dengan format upload GI Raw Material.<br/> Detail : " + ex.Message);
-                errorHandler.UploadCsvErrorHandler(loadCsvException, CsvDataGridView, errorLabel);
+                errorHandler.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
             }
         }
 
@@ -219,7 +220,7 @@ namespace KITE.Pages.ContentPages
                     Exception deleteResult = databaseModel.DeleteData(tableName, $" Year_Period = '{yearPeriod}' AND Month_Period = '{monthPeriod}' ", ConfigurationManager.ConnectionStrings["DefaultConnection"].ConnectionString);
                     if (deleteResult.Message != "Data period sebelumnya berhasil dibersihkan.")
                     {
-                        utility.UploadCsvErrorHandler(deleteResult, CsvDataGridView, errorLabel);
+                        utility.UploadCsvErrorHandler(deleteResult, new GridView[] { CsvDataGridView }, errorLabel);
                     }
                 }
 
@@ -227,40 +228,42 @@ namespace KITE.Pages.ContentPages
                 if (checkPeriodResult.Message != "Data Period Aman.")
                 {
                     loadCsvException = new Exception(checkPeriodResult.Message);
-                    utility.UploadCsvErrorHandler(loadCsvException, CsvDataGridView, errorLabel);
+                    utility.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
                 }
 
                 Tuple<string[], Exception> iterationResult = new ReadCsvModel().IterateCsvObject(columnNameAndData.Item2);
                 if (iterationResult.Item2.Message != "Pemrosesan IterateCsvObject Berhasil.")
                 {
                     loadCsvException = new Exception($"Gagal dalam penulisan ke database. Detail : {iterationResult.Item2.Message}");
-                    utility.UploadCsvErrorHandler(loadCsvException, CsvDataGridView, errorLabel);
+                    utility.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
                 }
 
-                foreach (string iterionValue in iterationResult.Item1)
+                if (checkPeriodResult.Message == "Data Period Aman." && iterationResult.Item2.Message == "Pemrosesan IterateCsvObject Berhasil.")
                 {
-                    Exception insertResult = new DatabaseModel().InsertIntoTable(tableName, columnNameAndData.Item1, iterionValue, ConfigurationManager.ConnectionStrings["DefaultConnection"].ConnectionString);
-                    if (insertResult.Message != "null")
+                    foreach (string iterionValue in iterationResult.Item1)
                     {
-                        loadCsvException = new Exception($"Terjadi kesalahan ketika Insert Into Table {tableName}. Detail : {insertResult.Message}");
-                        utility.UploadCsvErrorHandler(loadCsvException, CsvDataGridView, errorLabel);
-                        break;
+                        Exception insertResult = new DatabaseModel().InsertIntoTable(tableName, columnNameAndData.Item1, iterionValue, ConfigurationManager.ConnectionStrings["DefaultConnection"].ConnectionString);
+                        if (insertResult.Message != "null")
+                        {
+                            loadCsvException = new Exception($"Terjadi kesalahan ketika Insert Into Table {tableName}. Detail : {insertResult.Message}");
+                            utility.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
+                            break;
+                        }
                     }
-                }
 
-                if (File.Exists(Session["FilePath"].ToString()))
-                {
-                    File.Delete(Session["FilePath"].ToString());
+                    if (File.Exists(Session["FilePath"].ToString()))
+                    {
+                        File.Delete(Session["FilePath"].ToString());
+                    }
+                    Session["FilePath"] = "";
+                    string script = $"alert('Upload CSV berhasil.'); window.location.href = '{ResolveUrl("~/Pages/ContentPages/GIRawMaterialUpload.aspx")}';";
+                    ClientScript.RegisterStartupScript(this.GetType(), "SuccessAlert", script, true);
                 }
-                Session["FilePath"] = "";
-                string script = $"alert('Upload CSV berhasil.'); window.location.href = '{ResolveUrl("~/Pages/ContentPages/GIRawMaterialUpload.aspx")}';";
-                ClientScript.RegisterStartupScript(this.GetType(), "SuccessAlert", script, true);
-
             }
             catch (Exception ex)
             {
                 loadCsvException = new Exception(ex.Message);
-                utility.UploadCsvErrorHandler(loadCsvException, CsvDataGridView, errorLabel);
+                utility.UploadCsvErrorHandler(loadCsvException, new GridView[] { CsvDataGridView }, errorLabel);
             }
         }
     }
